// packages/bits/src/lib/panel/public-api.ts
export enum PanelModes {
    static = "static",
    collapsible = "collapsible",
    closable = "closable",
    hoverable = "hoverable",
}

export enum PanelStates {
    opened = "opened",
    collapsed = "collapsed",
}

export enum SizeParameters {
    width = "width",
    height = "height",
}

export enum PanelBackgroundColor {
    colorBgLight = "color-bg-light",
    colorBgSecondary = "color-bg-secondary",
    colorBgTransparent = "color-bg-transparent",
}